1.0

+++ src/main/java/com/AtpTournament/TennisAtpTournament/service/PlayerService.java
@@ -2,6 +2,7 @@ package com.AtpTournament.TennisAtpTournament.service;
 
 import com.AtpTournament.TennisAtpTournament.entity.Player;
 import com.AtpTournament.TennisAtpTournament.entityDto.PlayerDto;
+import com.AtpTournament.TennisAtpTournament.exception.InvalidPlayerException;
 import com.AtpTournament.TennisAtpTournament.exception.PlayerNotFoundException;
 import com.AtpTournament.TennisAtpTournament.mapper.PlayerMapper;
 import com.AtpTournament.TennisAtpTournament.repository.PlayerRepository;
@@ -12,7 +13,6 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 
 @Slf4j
 @Service
@@ -25,61 +25,94 @@ public class PlayerService {
     private PlayerMapper playerMapper;
 
     public int CreatePlayer(PlayerRequest playerRequest) {
+        log.info("Creating new player");
         Player player = playerMapper.PlayerRequestToPlayer(playerRequest);
+        if(player.getName() == null || player.getName().trim().isEmpty()) {
+            throw new InvalidPlayerException("Player name cannot be empty");
+        }
         playerRepository.save(player);
+        log.info("Player with name  " + playerRequest.getName() + "  created!");
         return player.getId();
     }
 
     public List<PlayerDto> GetPlayerByName(String name) {
+        log.info("Getting players by name...");
         List<PlayerDto> playersDto = new ArrayList<>();
         List<Player> players = playerRepository.findAllByNameIgnoreCase(name);
+        if(players.isEmpty()) {
+            throw new InvalidPlayerException("player with name  " + name + " not found");
+        }
         for (Player player : players) {
             playersDto.add(playerMapper.PlayerToPlayerDto(player));
         }
-
+        log.info("Player/s with name " + name +  " found!");
        return playersDto;
     }
 
     public PlayerDto GetPlayerById(Long id) {
-        log.info("Richiesta per ottenere il giocatore con ID: {}", id);
+        log.info("Searching player by id : {}", id);
         Player player = playerRepository.findById(id).
                 orElseThrow(() -> new PlayerNotFoundException(id));
-        log.info("Giocatore trovato, costruzione del DTO in corso...");
+        log.info("Player found! dto sended");
         return playerMapper.PlayerToPlayerDto(player);
     }
 
     public void DeletePlayer(Long id) {
-        playerRepository.deleteById(id);
+        log.info("Deleting player by id : {}", id);
+        if(playerRepository.existsById(id)) {
+            playerRepository.deleteById(id);
+            log.info("Player deleted!");
+        }
+        else{
+            throw new PlayerNotFoundException(id);
+        }
     }
 
     public List<PlayerDto> GetAllPlayers() {
+        log.info("Searching all players");
         List<Player> players = playerRepository.findAll();
+        if(players.isEmpty()) {
+            throw new InvalidPlayerException("No players found");
+        }
         List<PlayerDto> playersDto = new ArrayList<>();
         for (Player player : players) {
             playersDto.add(playerMapper.PlayerToPlayerDto(player));
         }
+        log.info("All player provided!");
         return playersDto;
     }
 
     public List<PlayerDto> GetAllPlayersBySponsor(String sponsor) {
+        log.info("Searching all players by sponsor : {}", sponsor);
         List<PlayerDto> playersDto = new ArrayList<>();
         List<Player> players = playerRepository.findAllBySponsorIgnoreCase(sponsor);
+        if(players.isEmpty()) {
+            throw new InvalidPlayerException("Sponsor not found");
+        }
         for (Player player : players) {
             playersDto.add(playerMapper.PlayerToPlayerDto(player));
         }
+        log.info("Plyers sorted by sponsor " + sponsor + " provided!");
         return playersDto;
     }
 
     public PlayerDto UpdatePlayer(Long id, PlayerRequest playerRequest) {
+        log.info("Updating player with id : {}", id);
         Player player = playerRepository.findById(id).
-                orElseThrow(() -> new IllegalArgumentException("Player not found by id" + id));
+                orElseThrow(() -> new PlayerNotFoundException(id));
         playerMapper.updatePlayer(playerRequest, player);
         playerRepository.save(player);
+        log.info("Player updated!");
         return playerMapper.PlayerToPlayerDto(player);
     }
 
     public List<PlayerDto> getPlayerRanking() {
+        log.info("filter players by ranking");
         List<Player> players = playerRepository.findAllByOrderByRankingAtpDesc();
+        if(players.isEmpty()) {
+            throw new InvalidPlayerException("No players found");
+        }
+        log.info("players filtered!");
         return players.stream()
                 .map(playerMapper::PlayerToPlayerDto)
                 .toList();
